Fix get reports by date

+++ api/app/repository/get_statistics.go
@@ -58,10 +58,10 @@ const (
 	order by date desc limit 1`
 
 	sqlGetStatistics = `
-	select avg(work_time) as avg_hours_worked,   -- среднне количество отработанных часов
-       	   sum(work_time) as hours_worked,       -- отработаннно часов
-       	   avg(break_time) as avg_hours_break,   -- среднне количество часов перерыва
-           avg(start_time) as avg_start_time     -- среднее время начала дня
+	select avg(work_time)::bigint as avg_hours_worked,   -- среднне количество отработанных часов
+       	   sum(work_time)::bigint as hours_worked,       -- отработаннно часов
+       	   avg(break_time)::bigint as avg_hours_break,   -- среднне количество часов перерыва
+           avg(start_time)::bigint as avg_start_time     -- среднее время начала дня
 	from main.reports
 	inner join main.reports_to_users rtu on reports.id = rtu.report_id
 	where rtu.user_id = $1 and
